[refactor] Extract plot utils into dedicated file (#1000)
* allocate helper functions from all 4 plotting .py files in utils_plot.py

* reallocate plotting-related helper functions from utils.py to utils_plot.py

* adapt import statements and reformate

* adapt import statement

* rename plot_utils.py
